refactor(server): setup db script
