+++ packages/twenty-server/src/database/scripts/truncate-db.ts
@@ -1,6 +1,6 @@
 import { rawDataSource } from 'src/database/typeorm/raw/raw.datasource';
 
-import { performQuery } from './utils';
+import { performQuery } from './setup-db-utils';
 
 async function dropSchemasSequentially() {
   try {
